fix: stabilize PR 1494 CI

+++ db/migrate/20260404203000_add_polymorphic_participant_to_agreement_participants.rb
@@ -4,8 +4,11 @@ class AddPolymorphicParticipantToAgreementParticipants < ActiveRecord::Migration
   disable_ddl_transaction!
 
   def up
-    add_column :better_together_agreement_participants, :participant_type, :string unless column_exists?(:better_together_agreement_participants, :participant_type)
-    add_column :better_together_agreement_participants, :participant_id, :uuid unless column_exists?(:better_together_agreement_participants, :participant_id)
+    add_column :better_together_agreement_participants, :participant_type, :string unless column_exists?(
+      :better_together_agreement_participants, :participant_type
+    )
+    add_column :better_together_agreement_participants, :participant_id, :uuid unless column_exists?(:better_together_agreement_participants,
+                                                                                                     :participant_id)
 
     execute <<~SQL.squish
       UPDATE better_together_agreement_participants
@@ -16,21 +19,28 @@ class AddPolymorphicParticipantToAgreementParticipants < ActiveRecord::Migration
     change_column_null :better_together_agreement_participants, :participant_type, false
     change_column_null :better_together_agreement_participants, :participant_id, false
 
-    add_index :better_together_agreement_participants,
-              %i[agreement_id participant_type participant_id],
-              unique: true,
-              algorithm: :concurrently,
-              name: 'idx_better_together_agreement_participants_unique_participant' unless index_exists?(:better_together_agreement_participants,
-                                                                                                         %i[agreement_id participant_type participant_id],
-                                                                                                         unique: true,
-                                                                                                         name: 'idx_better_together_agreement_participants_unique_participant')
+    unless index_exists?(:better_together_agreement_participants,
+                         %i[agreement_id participant_type
+                            participant_id],
+                         unique: true,
+                         name: 'idx_better_together_agreement_participants_unique_participant')
+      add_index :better_together_agreement_participants,
+                %i[agreement_id participant_type participant_id],
+                unique: true,
+                algorithm: :concurrently,
+                name: 'idx_better_together_agreement_participants_unique_participant'
+    end
   end
 
   def down
-    remove_index :better_together_agreement_participants,
-                 name: 'idx_better_together_agreement_participants_unique_participant' if index_exists?(:better_together_agreement_participants,
-                                                                                                        name: 'idx_better_together_agreement_participants_unique_participant')
-    remove_column :better_together_agreement_participants, :participant_type if column_exists?(:better_together_agreement_participants, :participant_type)
-    remove_column :better_together_agreement_participants, :participant_id if column_exists?(:better_together_agreement_participants, :participant_id)
+    if index_exists?(:better_together_agreement_participants,
+                     name: 'idx_better_together_agreement_participants_unique_participant')
+      remove_index :better_together_agreement_participants,
+                   name: 'idx_better_together_agreement_participants_unique_participant'
+    end
+    remove_column :better_together_agreement_participants, :participant_type if column_exists?(:better_together_agreement_participants,
+                                                                                               :participant_type)
+    remove_column :better_together_agreement_participants, :participant_id if column_exists?(:better_together_agreement_participants,
+                                                                                             :participant_id)
   end
 end
